fix: UInt32Array

+++ index.js
@@ -64,7 +64,7 @@ class Module {
                     }
                 }
                 else if (map == 'serpentine-matrix-x') {
-                    map = new Uint16Array(width*height);
+                    map = new Uint32Array(width*height);
                     for(var i = 0; i < map.length; i++) {
                         var row = Math.floor(i/height);
                         var col = height - (i % height) - 1;
@@ -74,7 +74,6 @@ class Module {
                         if (row % 2 === 1) {
                             target = width * col + row;
                         }
-
 
                         map[i] = target;
                     }
